+++ Bouncing Balls/Bouncing Balls/SettingsScene.h
@@ -7,7 +7,7 @@
 class SettingsScene : public Scene
 {
 public:
-	SettingsScene(uint8_t id, std::string name);
+	SettingsScene(int id, std::string name);
 	void init();
 	void update(sf::RenderWindow* w);
 	void input(sf::Event* e);
